chore: use pkg/manager instead of internal/manager everywhere
