+++ internal/cmd/rootcmd/run.go
@@ -8,9 +8,9 @@ import (
 
 	"sigs.k8s.io/controller-runtime/pkg/healthz"
 
-	"github.com/kong/kubernetes-ingress-controller/v3/internal/manager"
-	"github.com/kong/kubernetes-ingress-controller/v3/internal/manager/health"
+	"github.com/kong/kubernetes-ingress-controller/v3/pkg/manager"
 	managercfg "github.com/kong/kubernetes-ingress-controller/v3/pkg/manager/config"
+	"github.com/kong/kubernetes-ingress-controller/v3/pkg/manager/health"
 )
 
 // Run sets up a default stderr logger and starts the controller manager.
@@ -26,7 +26,11 @@ func Run(ctx context.Context, c managercfg.Config, output io.Writer) error {
 	}
 	defer signal.Ignore(shutdownSignals...)
 
-	m, err := manager.New(ctx, c, logger)
+	mid, err := manager.NewID("kic")
+	if err != nil {
+		return fmt.Errorf("failed to create manager ID: %w", err)
+	}
+	m, err := manager.NewManager(ctx, mid, logger, c)
 	if err != nil {
 		return fmt.Errorf("failed to create manager: %w", err)
 	}
