+++ test/envtest/run.go
@@ -16,9 +16,9 @@ import (
 	"k8s.io/client-go/rest"
 	"sigs.k8s.io/controller-runtime/pkg/healthz"
 
-	"github.com/kong/kubernetes-ingress-controller/v3/internal/manager/health"
 	"github.com/kong/kubernetes-ingress-controller/v3/pkg/manager"
 	managercfg "github.com/kong/kubernetes-ingress-controller/v3/pkg/manager/config"
+	"github.com/kong/kubernetes-ingress-controller/v3/pkg/manager/health"
 	"github.com/kong/kubernetes-ingress-controller/v3/test/mocks"
 )
 
@@ -230,7 +230,10 @@ func RunManager(
 		mgrID, err := manager.NewID(t.Name())
 		require.NoError(t, err)
 
-		mgr, err := manager.NewManager(ctx, mgrID, logger, modifyCfgFns...)
+		cfg, err := manager.NewConfig(modifyCfgFns...)
+		require.NoError(t, err)
+
+		mgr, err := manager.NewManager(ctx, mgrID, logger, cfg)
 		require.NoError(t, err)
 
 		if mgr.Config().ProbeAddr != "" {
