+++ test/envtest/setup.go
@@ -30,8 +30,8 @@ import (
 	"github.com/kong/kubernetes-ingress-controller/v3/internal/gatewayapi"
 	"github.com/kong/kubernetes-ingress-controller/v3/internal/store"
 	"github.com/kong/kubernetes-ingress-controller/v3/internal/util/builder"
-	testutils "github.com/kong/kubernetes-ingress-controller/v3/internal/util/test"
 	"github.com/kong/kubernetes-ingress-controller/v3/test/consts"
+	testutils "github.com/kong/kubernetes-ingress-controller/v3/test/util"
 )
 
 type Options struct {
