+++ test/util/crds.go
@@ -1,4 +1,4 @@
-package test
+package util
 
 import (
 	"context"
