+++ test/util/gomod.go
@@ -1,4 +1,4 @@
-package test
+package util
 
 import (
 	"fmt"
