+++ test/util/name.go
@@ -1,4 +1,4 @@
-package test
+package util
 
 import "github.com/google/uuid"
